Complete the WASM/native binding: echo, native .so, sim-fatal, determinism
- PAR-01: echo.nim reference vignette; Nim→WASM echo byte-matches TS.
- ABI-18: SimFatalError marker on the ABI. A WASM failure (nonzero return or
  trap) in handleMessage is sim-fatal, not peer-fault — WasmVignette poisons the
  instance and throws SimFatalError; the host routes it to shutdown. Verified TS
  (conformance) and WASM (faulty.nim).
- PAR-05: wg_vf.h offsets are uintptr_t (32-bit wasm / 64-bit native) — fixes
  pointer truncation on native. Compiles clean -Wall -Wextra -std=c11. The same
  counter.nim built as a native .so byte-matches the TS counter through a
  bun:ffi harness reading the ring buffer + frame from native memory.
- DET (crown jewel): lossyPipe (T-LOSSY); DET-03 counter TS vs WASM produce
  byte-identical frame+App traces through the host; DET-04 frame loss changes
  only received frames, App/event stream intact, sim unaffected.

Builds: npm run test:wasm:build / test:native:build. Artifacts gitignored
(.nimcache*). Tests: 71 pass across 10 files.

+++ src/vignettes/wasm/wg_vf.h
@@ -11,6 +11,10 @@
  * to the sim — a trapped instance's memory is not trustworthy (§2.4).
  *
  * Byte order is little-endian throughout, matching the wire envelope.
+ *
+ * Pointers/offsets are uintptr_t: a 32-bit linear-memory offset on wasm32, a
+ * native pointer on a 64-bit host. Everything else (ids, lengths, step values,
+ * return codes) is uint32_t on every target.
  */
 
 #ifndef WG_VF_H
@@ -25,7 +29,7 @@ extern "C" {
 /* --- Lifecycle (host-driven, serial; §2.2) ------------------------------- */
 
 /* Once, before anything else. Payload staged at (in_ptr, in_len). */
-uint32_t vf_init(uint32_t in_ptr, uint32_t in_len);
+uint32_t vf_init(uintptr_t in_ptr, uint32_t in_len);
 
 /* Once per host loop iteration while READY. */
 uint32_t vf_tick(uint32_t dt_us, uint32_t frame_id);
@@ -34,7 +38,7 @@ uint32_t vf_tick(uint32_t dt_us, uint32_t frame_id);
 uint32_t vf_fixed_tick(uint32_t step_us, uint32_t step_index);
 
 /* Per inbound App envelope. sender_id is host-stamped (§1.3). */
-uint32_t vf_handle_message(uint32_t sender_id, uint32_t in_ptr, uint32_t in_len);
+uint32_t vf_handle_message(uint32_t sender_id, uintptr_t in_ptr, uint32_t in_len);
 
 /* After a peer joins, before its first vf_handle_message. */
 uint32_t vf_peer_joined(uint32_t client_id);
@@ -52,8 +56,8 @@ uint32_t vf_shutdown(void);
  * Each entry, starting at offset 12: [payload_len u32][target_id u16][payload].
  * target_id 0 = broadcast; nonzero = unicast to that peer (§1.3).
  */
-uint32_t vf_outbox_offset(void);
-uint32_t vf_outbox_capacity(void);
+uintptr_t vf_outbox_offset(void);
+uint32_t  vf_outbox_capacity(void);
 
 /* --- Frame channel (latest-wins; §1.4) -----------------------------------
  *
@@ -61,14 +65,14 @@ uint32_t vf_outbox_capacity(void);
  * fixedTick burst and stamps the frame with the sim step. vf_frame_seq() is the
  * vignette-owned monotonic frameSeq. A length of 0 means "no frame yet".
  */
-uint32_t vf_frame_offset(void);
-uint32_t vf_frame_len(void);
-uint32_t vf_frame_seq(void);
+uintptr_t vf_frame_offset(void);
+uint32_t  vf_frame_len(void);
+uint32_t  vf_frame_seq(void);
 
 /* --- Input staging (host writes payloads here before calls) --------------- */
 
-uint32_t vf_mem_alloc(uint32_t size);
-void     vf_mem_free(uint32_t ptr);
+uintptr_t vf_mem_alloc(uint32_t size);
+void      vf_mem_free(uintptr_t ptr);
 
 #ifdef __cplusplus
 }
